+++ includes/fractol_image.h
@@ -35,7 +35,7 @@ typedef struct s_line
 }	t_line;
 
 t_line	line(size_t x, size_t y, size_t size, int color);
-void	my_mlx_pixel_put(t_img *img, int x, int y, int color);
+void	my_mlx_pixel_put(t_img *img, size_t x, size_t y, int color);
 void	cpy_img(t_img *dest, const t_img *src);
 void	img_put_line(t_img *img, t_line line);
